frontend/scripts/requirements.txt
google-genai>=1.70.0
google-cloud-storage>=2.18.0
